Dump is complete taking 22 minutes and 31 seconds with upping the provisioning to 400 during the process. Everything should be inserted, but I need to check the counts tomorrow. I expect 288,106 records.
I noticed an issue with certain characters being malformed. For example e acute is coming through oddly. https://en.wikipedia.org/wiki/%C3%89

I'm not sure what is the best approach. Either re-scan and update the classification or do updates within dynamo db. Or fix the issue and just do a complete re-classification. Do it without dumping the images, just get the raw html, classify it, and take in all fixes made up to this point. That would address the issue of the dynamo db backup.

Here's an example of the issue: http://www.the-athenaeum.org/art/detail.php?ID=165488

The artists name is: Léon Bakst

The classification file and dynamodb has: LÃ©on Bakst

It's probably simplest to do a search and replace, but that means the process is getting more complex. I'll do a combination. I'll search for the malformed characters, get new classifications, then update directly into dynamo db.

So the next phase is:

1. Find all page id's with the characters: Ã©
2. Re-classify all of the pages from the site directly into dynamodb with the correct name.

I checked and the characters are wrong on the raw hml file, so I need to investigate why the accent characters aren't being pulled correctly.