Fix E accute accent character in dynamodb and all utf-8 characters in the data dump. 3 records were lost in the process in dynamodb during the process. The classification is now done transiently directly against the active site.
There are several issues which exist:
 - It's possible that more pages have been added. The max page can easily be checked.
 - There are 3 known records missing (this is a bigger issue of integrity and backups)
 - There are additional utf-8 characters that are still incorrect.

It may be best to re-classify transiently over the throttled connection once I get an interface for the data established and get stable. I can download the images, only if they don't already exist, because I still want backups. That should cut the index time in half. I have a crawler and persistence so why not re-index? It's just a time issue. The only reason not to re-index is because there is no backup mechanism. Dynamodb has something to handle this, but it appears somewhat involved.

There is also the billing issue. I don't know what to expect as far as billing. When importing and classifying transiently the estimated cost is at least $100 per month, due to increasing the capacity, but this is likely if I kept that rate 24/7. As of this moment the capacity is 5/5 estimated at only $2.91 per month.