It took me 5 hours to get the challenge question. I may not be done yet, but I've done so in a SAFE manner. CloudFlare has a lot of smart people and I've heard Matthew Prince speak of turning attackers against themselves (I don't even want to think of what could potentially be done if I was consuming anonymous JavaScript), so I'm highly cautious to simply compile the JavaScript and execute whatever is given to me. Instead, I've reverse engineered the obfuscation and decoding. It's actually really well done right now. The only issue from a readability standpoint is the nested parentheses were just removed rather than traversing those paths. I could see the encoding process get more complex. However if you consider this only took me 5 hours from the last commit at 9:00am, I can probably move faster than CloudFlare. Although I'm sure they have other techniques ready to be deployed.
I'm taking an educated guess that this is what the I'm under attack mode does. It provides a challenge question which even has a built-in 4 second timer before even posting, so no matter what you need to specifically target your efforts to CloudFlare even with a headless browser. Then you start to play with the devil. Even with a headless browser and not a raw javascript engine, do you eval unkown javascript after stripping out the JavsScript wait? I would still be cautious. At most I'd go so far as executing the math operations, but even then you have to validate. I may see what get's done, but at the end of the day you have to allow an anonymous user access or it's not public. I'm surprised I was able to move through this so quickly. I thought this might take me a full weekend or more. Let's see if this is all that is needed.