Going to try and get some sleep and I will start enumerating tomorrow. Now I have a very cool idea. Once I enumerate the national gallery of art, I will then pull the images for the-athenaeum into s3. Now I'm thinking at this point I will classify them all. Eghh, Idk that's going to cost $350. Is it worth it? In the name of data science? Has anyone yet to analyze 350,000 pieces of fine art? Maybe one organization has artsy e.g. I'm pretty sure this is computer generated => https://www.artsy.net/gene/nude. Now, the fact that I'm going to compete with these guys in my spare time, just might be worth $350. It's actually not a competition, though they are concerned about selling new art. I on the other hand am concerned about exploring fine art. I want a real pandora for art. Not their BS product => https://www.artsy.net/categories. But who would buy new art at several thousand dollars (unless if you are looking for an alternate currency) if you can buy a reproduction for $250 (print) to $1000 (hand-painted)? Not me that's for sure. I'll give it to artsy they have a lot, but I don't see classics. And come on this shows up as a landscape: https://www.artsy.net/gene/landscapes has https://www.artsy.net/artwork/joe-wardwell-go-ahead-and-jump. I think I am in a different problem arena than artsy altogether, but it goes to show, who's doing this type of cutting edge work for the free classics vs commercial enterprises for new art that most people will not care about in a hundred years?
The CompareFaces operation and the two search faces operations differ as follows:

The CompareFaces operation compares a face in a source image with faces in the target image. The scope of this comparison is limited to the faces detected in the target image. For more information, see Comparing Faces.

SearchFaces and SearchFacesByImage compare a face (identified either by a FaceId or an input image) with all faces in a given face collection. Therefore, the scope of this search is much larger. Also, because the facial feature information is persisted for faces already stored in the face collection, you can search for matching faces multiple times.

I think this is how I will approach AWS Rekognition:
1. http://docs.aws.amazon.com/rekognition/latest/dg/howitworks-labeling.html
2. http://docs.aws.amazon.com/rekognition/latest/dg/image-moderation.html
3. http://docs.aws.amazon.com/rekognition/latest/dg/howitworks-comparefaces.html

Idk, I'll figure out exactly where to take this project soon. I may take a break and figure out how to avoid paying $10 a month for NEST's subscription to get more than 3hrs of recorded data. That seems too easy though...I mean look at what I'm doing here...Once I throw in image analysis it will be pretty darn intense. Maybe "intelligent" crawling. I'm just not sure with how expensive image analysis appears to be.