Make index process generic. Introuduce a file queue indexer to manage the IIndex interface. It should be possible to scale and distribute the process now. For example, one entity can fetch items for indexing crawling doing light work, then another can do heavy processing based on the ID's. Especially important if we want to do image analysis and the process becomes:  1. Get html  2. Get image  3. Push to dynamo db  4. Push to s3  5. Call AWS rekognition for analysis
This process is going to get heavy and I think I like the idea of ripping through the site to get the items of interest then taking my time classifying/indexing those items. So there could potentially be:

 1. Crawler
 2. Processor

In this architecture one entity is responsible for finding things of interest and sending them to a queue. Then another entity is responsible for getting all related items for said item of interest and storing it for deep analysis: application with dynamo db/s3 or AWS rekognition.

Finally the Gallery will be the interface into this data.

 1. Crawlers
 2. Analyzers
 3. Interface

So far I've been doing the crawling from unit tests. For the national gallery of art I actually crawled the search page to get the ID's, because only a a percentage of the results had images. For the-athenaeum I just did a for loop. I can imagine these tasks being run on a schedule every so often. Then having a farm of analyzers picking up whatever the crawlers throw at it and handling the id based on source. This is important, because it's more work to analyze the ID's than to pull them so I don't want bottle necks in any process. By eliminating bottle necks I'm going to highlight true bottlenecks, which should be clear now that it's the FileIdQueueIndexer not getting the ID's.