Security is done in layers. Although I have gitignored many secrets. I assume it's already going to be stolen. So I encrypt it if that becomes the case. I'll share everything, but the raw secrets. There is a highly secretive website that had "an" IP address in source code. We must learn from such actions, even if they are decoys and never even let the QUESTION be raised.
Eventually I want an alarming system. So that on commit, I check for the contents of the personal.json which is expected to be gitignored in github (AWS has such alarms for their customers API keys). Then my logic is public so it can never be lost and if secrets are revealed, then only people such as myself with the diligence I am undergoing will ever be able to even begin to posses it. Now it does raise the question of what has been revealed on the open web? What will be? We will not know unless if we crawl ourselves.