Every backpage ad for persons under 30 in the united states on each front page of the geographic region links into dynamodb. This took 43 minutes. So 15 minutes to get all the links and 28 minutes to insert. That time could be cut down by a factor of 10 with bulk inserts to a mere 2.8 minutes. Meaning the entire process can run in under 20 minutes!
    Indexing ads for persons aged thirty and under:
    36218

Wait till I show tomorrow how I am going to distribute the work to pull back the content. It will be very impressive I promise. It might take a few days though. Crap there is one problem, to do the analysis I need java...Not just java, but java 7. Well technically I still need to work on pulling back the ad content, then I can deal with processing in Java. Now there is gate cloud. It's .8 GBP per cpu hour and I actually estimate that I will need 1 cpu hour (at least locally) per day. That's one dollar per day.  Ugh, I still don't know exactly what I'm going to do when I process the data. I mean what's the end goal here? Just build a data pipeline? The most interesting piece of data, price doesn't seem to even exist on most ads that I check. All of the ages are likely a lie. Same for the pictures. Phone number would be great to cross reference and extract, but where do I go and look? How? Google isn't indexing this stuff and most of it appears to be straight disinformation.

You know what would be insanely cool. A chat bot! What would be funny is a chat bot that tried to talk girls into getting help. Funny only in that they would ignore all advice. It would really just be one step down from TDOS.

Now here is something to look into, but it requires a lot of work doing automatic searches in google: search all phone numbers and seek out new websites. Then compile a list of all websites, crawl those, then keep at it over and over.

Really I would see the ultimate goal as finding the next backpage or crawling a new website with no manual intervention. On the point of fake pictures, perhaps the picture is always real, but simply old or re-used. Just browsing the site you can tell some pictures are definitely for prostitutes and some iffy. For example a picture of a girl in a cheap hotel room is 95% likely to be real and a prostitute. Now is that actually the entity for the phone number, who knows. It depends if the picture is new and unique. Perhaps, just perhaps if there was enough breadth in the data and there was an image of a very young girl in a cheap hotel, that would be in line with memex's mission statement, but who would be that stupid? How would it even be known from an image? I need to think more about this problem area. It almost seems useless unless if I was looking for a specific person.

Unless if there was no one-shot answer. Now listen to this, what if there was a picture of a young girl in a cheap hotel room that appeared across many geographic locations. Perhaps I need to connect dots like Emily Kennedy's paper spoke of. This is much harder analysis, but certainly possible with the types of daily scans I intend to do. But you see here's where it gets difficult, the phone number might change as the entity moves in which case I would have to rely on the image and it's expensive to do image analysis on 30,000+ images per day. It would cost $30 per day or $900 per month just do do one analysis @ $1 for 1,000 images. I don't know. I need to put a lot of though into this to actually pull out valuable information of some kind.

Well here's an easy task to start, try and estimate the total number of sex workers in each city. If you can even begin to estimate a number then you are identifying entities and you can track shifts and perhaps eventually move to a shifting single entity. None of this is easy though.

Summary of ideas:
 - Chatbot to communicate over advertised phone number
 - Seek out new sex websites with phone number
 - Attempt to identify entities with the ultimate goal of finding a particular one mover over a geographic region
 - Attempt to identify scenery in an attempt to identify images which are more indicative of sex workers vs those that could exist on any sex related website

Let's have this goal, which I can accomplish with just what I'm indexing now.

1. what are the number of unique ads based on geographic region. Which region has the most. which region has the least.

Next extract phone number from the ads. Then track the usage of a phone number across geographic regions. This is actually very simple, because I can just build an index on phone number, then cross reference each ad with the phone numbers in all of the other ads. The url states the location, so any results I can order by ate, the url will have the location. This would actually be very informative. I mean it's 30,000 ads. People will change phone numbers, but not everyone. Wow, this is actually going to be pretty cool. I can track their movement. Maybe there will be something intersting there, maybe there will not be many relationships and re-used phone numbers. It is somewhat interesting and requires very little data points.