Rebuilt elastic search instance with a 4GB disk and 2GB memory optimized instance this time at a slightly higher price. Something was terribly wrong with the old instance and it couldn't be recovered. JVM memory pressure now sits at 20%. Shards were also decreased on the classification index from 3 to 1 since the data is only 200mb right now and will likely never reach 30gb or more or that would be tens of millions of classified images.
Starting to move out images in review so the images can't be downloaded. Went this route so one day the s3 bucket can be delivered right to users to serve images over 5.5MB which would otherwise be difficult with any type of hit to the db to check if the image was in review.

Adding pagination now. Pagination will now double the max search results from 5,000 to the built in search result limit of 10,000. This will be very helpful for allowing users to play a gallery on a screen to last for 16 hours. There is still a ways to go for multiple days, but I don't have such a search interface yet e.g. date range where I can reasonably target records in the range greater than 10,000.